+++ runtimes/cpp/core/GdbStub.cpp
@@ -63,11 +63,11 @@ int GdbStub::hexToNum(char c) {
 }
 
 void GdbStub::checkAndResize(int len) {
-	unsigned long curIndex = (unsigned long)(curOutputBuffer-outputBuffer.begin());
-	unsigned long newSize = (len+curIndex)+1; // +1 for the null terminator...
+	size_t curIndex = (size_t)(curOutputBuffer-outputBuffer.begin());
+	size_t newSize = (len+curIndex)+1; // +1 for the null terminator...
 	if(newSize>(outputBuffer.size())) {
 		//BIG_PHAT_ERROR(ERR_INTERNAL);
-		unsigned long betterSize = (outputBuffer.size()+2)*2;
+		size_t betterSize = (outputBuffer.size()+2)*2;
 		if(betterSize>newSize) newSize=betterSize;
 		outputBuffer.resize(newSize);
 		curOutputBuffer = outputBuffer.begin()+curIndex;
